requirements.txt
asgiref==3.9.1
Django==5.2.5
django-filter==25.1
djangorestframework==3.16.1
environs==14.3.0
Markdown==3.8.2
marshmallow==4.0.0
psycopg2==2.9.10
python-dotenv==1.1.1
sqlparse==0.5.3
tzdata==2025.2
